refactor: rename file "claimUserTickets" to "claimUserTicket"
