+++ src/schema/userTickets/mutations/claimUserTicket.ts
